move and split abstract.py into object.py and tensor.py
